+++ CHANGELOG.md
@@ -34,3 +34,9 @@
 
 ## Context Integrated
 - Creación de *CustomProvider*, componente que crea y provee el contexto a los demás componentes del estado de compra del Usuario.
+
+## Cart Component Developed
+- Desglose de productos de productos elegidos por el usuario
+- Provisión de controles para eliminar productos del carrito
+- Muestra del total a pagar por el usuario en base a productos seleccionados (Nueva función getTotalPrice en CartContext)
+- Se arregló CartWidget (Dirige a /cart y no muestra "0" cuando no hay productos seleccionados)